Set properties for VM to clone
Proxmox and CAPMOX require and initial VM to clone. This VM does not
need to have booted before. But it needs to be present and configured
with the correct ISO image

Note, the Guest agent also needs to be enabled in this VM template as we
will be using the Qemu Guest Agent in Talos to communicate the node's
status back to Proxmox and the infrastructure provider within CAPI
(CAPMOX)