Set Virtual IP for control plane using known interface name
Traditionally, an external load balancer would be sat in front of
control plane nodes and distribute traffic destined to the control plane
between nodes. However, that adds another layer of complexity. Here, we
are making use of Talos's in-built VIP (Virtual IP) functionality. A
node gets issued with the control plane node and all traffic to the
control plane will go to this node. If this node goes down, the control
plane IP gets allocated to another CP node that is healthy. This
abstracts the control plane to a single IP. However, this isn't true
load-balancing