Add Canal manifests to Talos installation
The default CNI that ships with Talos (Flannel) does not support
Kubernetes NetworkPolicies to control what network traffic that pods can
receive or transmit traffic to/from. There are more advanced CNIs like
Calico/Cillium etc that do support NetworkPolicies. However, this adds
extra complication with the installation as, for Cillum, the
installation either requires the `cillium` CLI, or Helm. Installing via
rendered manifests is also possible, but the data in the manifests is
deemed sensitive so the Talos machine config patches should not be
stored in source control

A compromise here is to use Canal (Calico + Flannel). This is the
Flannel CNI, which is installed in Talos by default. But also uses
Calico in a "NetworkPolicies mode" which is used only to enforce
NetworkPolicy rules defined in application manifests

Whilst this doesn't come with benefits of a more feature-rich CNI such
as Cillium, which as the ability to perform service Load balancing
(negating the need for something like MetalLB), the installation/setup
of Canal keeps the Talos installation via CAPI simple