Add DNS of gateway (router) to nodes
To allow the Kubernetes nodes to resolve ingress/HTTPRoutes etc that
have been created by the _same_ cluster, the nodes need to have it using
the DNS server of the UniFi gateway

The use case for this is to perform health checks with a tool like
Gatus, to make periodic requests to URLs of ingress/Gateway API
resources to ensure the HTTP response is valid and the application is
up/healthy. In an ideal world, monitoring/observability tooling would
not be installed on the same cluster as the one it monitors. One, it
compounds the observer effect. But two, if the cluster goes down, so
does all monitoring tooling. I.e. you have no way of being alerted that
there's an issue! However, seeing as this is simply a homelab for
learning, I can live with that